src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pypdf2==3.0.1
chromadb==0.4.18
openai==1.3.0
python-multipart==0.0.6
langchain==0.0.340
tiktoken==0.5.1
cors==1.0.1
fastapi-cors==0.0.6
python-docx==1.1.0

